requirements.txt
Django==5.0.2
python-dotenv==1.0.1
requests==2.31.0
djangorestframework==3.14.0 